tenancy done - existing cmds wip
