--- stubs/method.stub
@@ -1,4 +1,0 @@
-    public function {{ name }}(): {{ returnType }}
-    {
-        return $this->{{ relationshipName }}({{ $modelClass }});
-    }
